% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esq_handsontable.R
\docType{package}
\name{esq.handsontable-package}
\alias{esq.handsontable}
\alias{esq.handsontable-package}
\title{esq.handsontable: Interactive Data Tables with 'Handsontable' for 'Shiny'}
\description{
Provides an interactive, Excel-like data table input for 'Shiny' applications. Wraps 'Handsontable' \url{https://handsontable.com/} and 'React' to offer editable tables with dropdown columns, multi-select with sorting, validation, conditional cell disabling, context menu, and dynamic option updates at runtime.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/esqlabs/esq.handsontable}
  \item Report bugs at \url{https://github.com/esqlabs/esq.handsontable/issues}
}

}
\author{
\strong{Maintainer}: Anastasia Kostiv \email{anastasiia.kostiv.ext@esqlabs.com}

}
\keyword{internal}
